IOS-6324 Moved adaptor out of amount VM
