+++ Tangem/Modules/Send/SendViewModel.swift
@@ -93,6 +93,7 @@ final class SendViewModel: ObservableObject {
     private let emailDataProvider: EmailDataProvider
     private let walletInfo: SendWalletInfo
     private let notificationManager: CommonSendNotificationManager
+    private let fiatCryptoAdapter: CommonSendFiatCryptoAdapter
     private let sendStepParameters: SendStep.Parameters
 
     private weak var coordinator: SendRoutable?
@@ -195,9 +196,15 @@ final class SendViewModel: ObservableObject {
             input: sendModel
         )
 
+        fiatCryptoAdapter = CommonSendFiatCryptoAdapter(
+            cryptoCurrencyId: walletInfo.currencyId,
+            currencySymbol: walletInfo.cryptoCurrencyCode,
+            decimals: walletInfo.amountFractionDigits
+        )
+
         sendStepParameters = SendStep.Parameters(currencyName: walletModel.tokenItem.name, walletName: walletInfo.walletName)
 
-        sendAmountViewModel = SendAmountViewModel(input: sendModel, walletInfo: walletInfo)
+        sendAmountViewModel = SendAmountViewModel(input: sendModel, fiatCryptoAdapter: fiatCryptoAdapter, walletInfo: walletInfo)
         sendDestinationViewModel = SendDestinationViewModel(input: sendModel)
         sendFeeViewModel = SendFeeViewModel(input: sendModel, notificationManager: notificationManager, walletInfo: walletInfo)
         sendSummaryViewModel = SendSummaryViewModel(input: sendModel, notificationManager: notificationManager, walletInfo: walletInfo)
